Create Upload Music Service
Decouple Upload Music Service from Audio File Service. Fix CSS for webpage links
