+++ src/main/resources/templates/musicPage.html
@@ -14,12 +14,10 @@
 
 <div class="top-container">
   <a th:href="@{/}" class="home-link" >Home</a>
-  <a th:href="@{/upload}" class="another-link">Upload</a>
+  <a th:href="@{/upload}" class="upload-link">Upload</a>
 </div>
 
-<div>
-  <h2 class="title-40">Music Page</h2>
-</div>
+<h2 class="title-40">Music</h2>
 
 <div class="audio-container">
   <audio controls id="audio-player" preload="auto" mediagroup="audioGroup">
