+++ src/main/resources/templates/uploadMusicPage.html
@@ -15,10 +15,10 @@
 
 <div class="top-container">
   <a th:href="@{/}" class="home-link" >Home</a>
-  <a th:href="@{/music}" class="another-link">Music</a>
+  <a th:href="@{/music}" class="music-link">Music</a>
 </div>
 
-<h2 class="title-40">Upload Music Page</h2>
+<h2 class="title-40">Upload Music</h2>
 
 <div class="upload-container">
   <div class="form-group">
